+++ src/runtime/gecko/l20n-html.js
@@ -1,9 +1,14 @@
 import { prioritizeLocales } from '../../intl/locale';
-import { documentReady, getResourceLinks } from './util';
+import { valueFromContext } from '../../lib/format';
+
+import { ChromeLocalizationObserver } from '../../bindings/observer/chrome';
+import { HTMLLocalization, contexts } from '../../bindings/dom/html';
+
 import { ResourceBundle } from './resourcebundle';
-import { GeckoLocalization } from './localization';
+import { documentReady, getResourceLinks, observe } from './util';
 
 Components.utils.import('resource://gre/modules/Services.jsm');
+Components.utils.import('resource://gre/modules/IntlMessageContext.jsm');
 
 function requestBundles(requestedLangs = navigator.languages) {
   return documentReady().then(() => {
@@ -21,7 +26,38 @@ function requestBundles(requestedLangs = navigator.languages) {
   });
 }
 
-document.l10n = new GeckoLocalization(document, requestBundles);
+const functions = {
+  OS: function() {
+    switch (Services.appinfo.OS) {
+      case 'WINNT':
+        return 'win';
+      case 'Linux':
+        return 'lin';
+      case 'Darwin':
+        return 'mac';
+      case 'Android':
+        return 'android';
+      default:
+        return 'other';
+    }
+  }
+};
 
+function createContext(lang) {
+  return new MessageContext(lang, { functions });
+}
+
+const localization = new HTMLLocalization(requestBundles, createContext);
+localization.observe = observe;
+localization.interactive.then(bundles => {
+  localization.getValue = function(id, args) {
+    return valueFromContext(contexts.get(bundles[0]), id, args)[0];
+  };
+});
+
+Services.obs.addObserver(localization, 'language-update', false);
+
+document.l10n = new ChromeLocalizationObserver();
+document.l10n.observeRoot(document.documentElement, localization);
+document.l10n.translateRoot(document.documentElement);
 window.addEventListener('languagechange', document.l10n);
-Services.obs.addObserver(document.l10n, 'language-update', false);
