+++ src/runtime/testing/index.js
@@ -1,5 +1,2 @@
-import * as dom from '../../bindings/dom';
-
-export default {
-  dom
-};
+// XXX what needs to be here?
+export default {};
