+++ src/runtime/web/index.js
@@ -1,8 +1,9 @@
 import '../../intl/polyfill';
-
 import { prioritizeLocales } from '../../intl/locale';
+
 import { ContentLocalizationObserver } from '../../bindings/observer/content';
 import { HTMLLocalization } from '../../bindings/dom/html';
+
 import { ResourceBundle } from './resourcebundle';
 import { documentReady, getResourceLinks, getMeta } from './util';
 
@@ -26,6 +27,7 @@ function createContext(lang) {
 }
 
 document.l10n = new ContentLocalizationObserver();
+
 window.addEventListener('languagechange', document.l10n);
 
 document.l10n.observeRoot(
